scripts/requirements.txt
# Local PDF Processing Dependencies
openai>=1.3.0
pypdf>=3.17.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0

# poppler-utils gerekli (pdf2image için)
# macOS: brew install poppler
# Ubuntu: sudo apt-get install poppler-utils
# Windows: https://github.com/oschwartz10612/poppler-windows/releases/
